+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2025-09-23
+
+- Added CLI skeleton: `BackgammonCLI` with ASCII board header/footer rendering.
+- Exposed `BackgammonGame` and `BackgammonCLI` via package `__init__` files.
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -10,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Sprint 3
 
 ### [0.0.11] - 2025-09-23
+
+#### Added
+
+- skleton of cli
 
 #### Fix